+++ src/i18next.js
@@ -96,7 +96,6 @@ class I18n extends EventEmitter {
       } else if (typeof Intl !== 'undefined') {
         formatter = Formatter;
       }
-  
 
       const lu = new LanguageUtils(this.options);
       this.store = new ResourceStore(this.options.resources, this.options);
@@ -433,7 +432,7 @@ class I18n extends EventEmitter {
       return false;
     }
 
-    const lng = this.resolvedLanguage || this.languages[0];
+    const lng = options.lng || this.resolvedLanguage || this.languages[0];
     const fallbackLng = this.options ? this.options.fallbackLng : false;
     const lastLng = this.languages[this.languages.length - 1];